requirements.txt
python-binance==1.0.19
pandas==2.3.1
numpy==1.24.3
ta==0.11.0
aiohttp==3.12.15
websockets==15.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
scikit-learn==1.3.0
joblib==1.3.2
